requirements.txt
asgiref==3.8.1
boto3==1.34.95
botocore==1.34.95
Django==4.2.11
django-environ==0.11.2
django-filter==24.2
django-storages==1.14.2
djangorestframework==3.15.1
gunicorn==22.0.0
importlib_metadata==7.1.0
jmespath==1.0.1
Markdown==3.6
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==1.26.18
zipp==3.18.1
